feat: proof of concept
This provides proof of concept extensions for web3.py
to add Tempo support.

This works by monkeypatching `TypedTransaction` from `eth_account`,
which is used by `web3` internally.

Curently, this does not monkeypatch the fill + sign middleware,
as this mostly serves as an idea for what this would look like directionally.
As such, the library currently exposes a `create_tempo_transaction`, and the
examples rely on building, encoding and signing the transactions manually and in full.

If we replaced or monkeypatched the fill + sign middleware, this would look
even simpler, but the monkeypatch would not go away.